common methods to base pipeline
